# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guitar learning app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  teacher: "teacher",
  student: "student",
} as const;

export interface AdminLoginBody {
  password: string;
}

export interface CodeLoginBody {
  code: string;
  name: string;
}

export interface TeacherLoginBody {
  code: string;
  /** @minLength 2 */
  firstName: string;
  /** @minLength 2 */
  lastName: string;
}

export interface CheckCodeBody {
  code: string;
}

export type CodeCheckResultKind =
  (typeof CodeCheckResultKind)[keyof typeof CodeCheckResultKind];

export const CodeCheckResultKind = {
  teacher: "teacher",
  student: "student",
} as const;

export interface CodeCheckResult {
  kind: CodeCheckResultKind;
  institutionName: string;
  className?: string | null;
  used?: boolean;
}

export interface UserProfile {
  id: string;
  role: UserRole;
  name: string;
  firstName?: string | null;
  lastName?: string | null;
  institutionId?: string | null;
  institutionName?: string | null;
  classId?: string | null;
  className?: string | null;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface InviteCode {
  code: string;
}

export interface CreateInstitutionBody {
  name: string;
  /** @minimum 1 */
  teacherLimit: number;
  /** @minimum 1 */
  studentLimit: number;
}

export interface UpdateInstitutionLimitsBody {
  /** @minimum 1 */
  teacherLimit: number;
  /** @minimum 1 */
  studentLimit: number;
}

export interface TeacherCodeInfo {
  code: string;
  used: boolean;
}

export interface InstitutionWithStats {
  id: string;
  name: string;
  teacherLimit: number;
  studentLimit: number;
  totalTeachers: number;
  totalStudents: number;
  usedTeacherCount: number;
  usedStudentCount: number;
  remainingTeacherSlots: number;
  remainingStudentSlots: number;
  unusedTeacherCodes: number;
  teacherCodes: TeacherCodeInfo[];
}

export interface AdminStats {
  totalInstitutions: number;
  totalTeachers: number;
  totalStudents: number;
  totalClasses: number;
}

export interface CreateClassBody {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  studentCount: number;
}

export interface ExpandClassBody {
  /** @minimum 1 */
  additional: number;
}

export interface StudentCodeInfo {
  code: string;
  used: boolean;
  usedByName?: string | null;
}

export interface ClassWithStats {
  id: string;
  name: string;
  levelUnlocked: number;
  studentCount: number;
  studentCapacity: number;
  usedStudentCount: number;
  remainingSlots: number;
  unusedStudentCodes: number;
  studentCodes: StudentCodeInfo[];
}

export interface LessonWithProgress {
  id: string;
  moduleNumber: number;
  moduleTitle: string;
  code: string;
  title: string;
  description: string;
  level: number;
  locked: boolean;
  completed: boolean;
}

export interface StudentDashboard {
  studentName: string;
  className: string;
  currentLevel: number;
  completedLessons: number;
  totalLessons: number;
  nextLesson: LessonWithProgress | null;
}