# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Guitar learning app API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/admin-login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/check-code:
    post:
      operationId: checkInviteCode
      tags: [auth]
      summary: Inspect a teacher or student invite code without logging in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckCodeBody"
      responses:
        "200":
          description: Code info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CodeCheckResult"
        "404":
          description: Code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/teacher-login:
    post:
      operationId: teacherLogin
      tags: [auth]
      summary: Teacher login with code and identity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherLoginBody"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/student-login:
    post:
      operationId: studentLogin
      tags: [auth]
      summary: Student login with code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeLoginBody"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /admin/institutions:
    get:
      operationId: listInstitutions
      tags: [admin]
      summary: List all institutions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstitutionWithStats"
    post:
      operationId: createInstitution
      tags: [admin]
      summary: Create institution
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstitutionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstitutionWithStats"
  /admin/institutions/{id}:
    patch:
      operationId: updateInstitutionLimits
      tags: [admin]
      summary: Update teacher and student limits of an institution
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstitutionLimitsBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstitutionWithStats"
        "400":
          description: Invalid limits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/institutions/{id}/teacher-codes:
    post:
      operationId: generateTeacherCode
      tags: [admin]
      summary: Generate a teacher invite code for an institution
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteCode"
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Global stats for the admin dashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
...
  /student/lessons:
    get:
      operationId: listMyLessons
      tags: [student, lessons]
      summary: List lessons with progress and lock state for the current student
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LessonWithProgress"
  /student/lessons/{id}/complete:
    post:
      operationId: completeLesson
      tags: [student, lessons]
      summary: Mark a lesson as completed for the current student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonWithProgress"
  /student/dashboard:
    get:
      operationId: getStudentDashboard
      tags: [student]
      summary: Aggregated dashboard for the student home screen
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDashboard"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserRole:
      type: string
      enum: [admin, teacher, student]
    AdminLoginBody:
      type: object
      required: [password]
      properties:
        password:
          type: string
    CodeLoginBody:
      type: object
      required: [code, name]
      properties:
        code:
          type: string
        name:
          type: string
    TeacherLoginBody:
      type: object
      required: [code, firstName, lastName]
      properties:
        code:
          type: string
        firstName:
          type: string
          minLength: 2
        lastName:
          type: string
          minLength: 2
    CheckCodeBody:
      type: object
      required: [code]
      properties:
        code:
          type: string
    CodeCheckResult:
      type: object
      required: [kind, institutionName]
      properties:
        kind:
          type: string
          enum: [teacher, student]
        institutionName:
          type: string
        className:
          type: string
          nullable: true
        used:
          type: boolean
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
    UserProfile:
      type: object
      required: [id, role, name]
      properties:
        id:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        name:
          type: string
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        institutionId:
          type: string
          nullable: true
        institutionName:
          type: string
          nullable: true
        classId:
          type: string
          nullable: true
        className:
          type: string
          nullable: true
    InviteCode:
      type: object
      required: [code]
      properties:
        code:
          type: string
    CreateInstitutionBody:
      type: object
      required: [name, teacherLimit, studentLimit]
      properties:
        name:
          type: string
        teacherLimit:
          type: integer
          minimum: 1
        studentLimit:
          type: integer
          minimum: 1
    UpdateInstitutionLimitsBody:
      type: object
      required: [teacherLimit, studentLimit]
      properties:
        teacherLimit:
          type: integer
          minimum: 1
        studentLimit:
          type: integer
          minimum: 1
    InstitutionWithStats:
      type: object
      required: [id, name, teacherLimit, studentLimit, totalTeachers, totalStudents, usedTeacherCount, usedStudentCount, remainingTeacherSlots, remainingStudentSlots, unusedTeacherCodes, teacherCodes]
      properties:
        id:
          type: string
        name:
          type: string
        teacherLimit:
          type: integer
        studentLimit:
          type: integer
        totalTeachers:
          type: integer
        totalStudents:
          type: integer
        usedTeacherCount:
          type: integer
        usedStudentCount:
          type: integer
        remainingTeacherSlots:
          type: integer
        remainingStudentSlots:
          type: integer
        unusedTeacherCodes:
          type: integer
        teacherCodes:
          type: array
          items:
            $ref: "#/components/schemas/TeacherCodeInfo"
    TeacherCodeInfo:
      type: object
      required: [code, used]
      properties:
        code:
          type: string
        used:
          type: boolean
    AdminStats:
      type: object
      required: [totalInstitutions, totalTeachers, totalStudents, totalClasses]
      properties:
        totalInstitutions:
          type: integer
        totalTeachers:
          type: integer
        totalStudents:
          type: integer
        totalClasses:
          type: integer
...
    LessonWithProgress:
      type: object
      required: [id, moduleNumber, moduleTitle, code, title, description, level, locked, completed]
      properties:
        id:
          type: string
        moduleNumber:
          type: integer
        moduleTitle:
          type: string
        code:
          type: string
        title:
          type: string
        description:
          type: string
        level:
          type: integer
        locked:
          type: boolean
        completed:
          type: boolean
    StudentDashboard:
      type: object
      required: [studentName, className, currentLevel, completedLessons, totalLessons, nextLesson]
      properties:
        studentName:
          type: string
        className:
          type: string
        currentLevel:
          type: integer
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        nextLesson:
          $ref: "#/components/schemas/LessonWithProgress"
          nullable: true