# lib/api-zod/src/generated/types/codeCheckResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guitar learning app API
 * OpenAPI spec version: 0.1.0
 */
import type { CodeCheckResultKind } from "./codeCheckResultKind";

export interface CodeCheckResult {
  kind: CodeCheckResultKind;
  institutionName: string;
  className?: string | null;
  used?: boolean;
}