# lib/api-zod/src/generated/types/institutionWithStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guitar learning app API
 * OpenAPI spec version: 0.1.0
 */
import type { TeacherCodeInfo } from "./teacherCodeInfo";

export interface InstitutionWithStats {
  id: string;
  name: string;
  teacherLimit: number;
  studentLimit: number;
  totalTeachers: number;
  totalStudents: number;
  usedTeacherCount: number;
  usedStudentCount: number;
  remainingTeacherSlots: number;
  remainingStudentSlots: number;
  unusedTeacherCodes: number;
  teacherCodes: TeacherCodeInfo[];
}