# lib/api-zod/src/generated/types/teacherLoginBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guitar learning app API
 * OpenAPI spec version: 0.1.0
 */

export interface TeacherLoginBody {
  code: string;
  /** @minLength 2 */
  firstName: string;
  /** @minLength 2 */
  lastName: string;
}