# lib/api-zod/src/generated/types/updateInstitutionLimitsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guitar learning app API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateInstitutionLimitsBody {
  /** @minimum 1 */
  teacherLimit: number;
  /** @minimum 1 */
  studentLimit: number;
}